Add files via upload
This is a port of a perl prototype done quite some time ago by a friend that I'd like to give credit, but she'd probably rather remain anonymous. Anyway, most of the clever stuff is hers, the rest is mine and served as an introduction to python, tkinter, and object oriented programming. Unzip the card pic files into <progdir> and <progdir/resized> to play..